SYSTEM.h fix for darwin
Former-commit-id: 7ed322be7903745b420947387e58133e295a3578

+++ src/lib/system/darwin/clang/x86_64/SYSTEM.h
@@ -15,8 +15,7 @@ uses double # as concatenation operator
 //#include <string.h>
 
 extern void *memcpy(void *dest, const void *src, unsigned long n);
-//extern char *memcpy();
-extern void *malloc(unsigned long size);
+extern void *malloc(size_t size);
 extern void exit(int status);
 
 #define export
